Add DontCleanSymlinks option

+++ scripts/Setup-SymLinks.ps1
@@ -1,6 +1,7 @@
 [CmdletBinding()]
 param(
-	[String]$CustomRessourcePluginsPath=""
+	[String]$CustomRessourcePluginsPath="",
+	[Switch]$DontCleanSymlinks=$false
 )
 
 
@@ -53,20 +54,23 @@ Get-ChildItem -Path $RessourcesPluginsPath -Recurse -Filter "*.uplugin" | ForEac
 $RootPath = $InputUnrealProjectDirectoryPath
 
 if (-not (Test-Path $RootPath)) {
-        Write-Warning "Path does not exist: $RootPath"
-        return
-    }
-
-    Get-ChildItem -Path $RootPath -Recurse -Force | ForEach-Object {
-        if ($_.Attributes -band [IO.FileAttributes]::ReparsePoint) {
-            try {
-                Remove-Item $_.FullName -Force
-                Write-Host "Removed symlink: $($_.FullName)"
-            } catch {
-                Write-Warning "Failed to remove symlink: $($_.FullName)"
-            }
-        }
-    }
+	Write-Warning "Path does not exist: $RootPath"
+	return
+}
+		
+if(-not $DontCleanSymlinks)
+{
+	Get-ChildItem -Path $RootPath -Recurse -Force | ForEach-Object {
+		if ($_.Attributes -band [IO.FileAttributes]::ReparsePoint) {
+			try {
+				Remove-Item $_.FullName -Force
+				Write-Host "Removed symlink: $($_.FullName)"
+			} catch {
+				Write-Warning "Failed to remove symlink: $($_.FullName)"
+			}
+		}
+	}
+}
 
 foreach ($symlinkRelativePath in $SymLinks.Keys) 
 {
